+++ Fincompare.Application/Response/MerchantRemitProductRateResponse/MerchantRemitProductRateViewModel.cs
@@ -29,13 +29,13 @@
 
         public string ReceiveCur { get; set; }
 
-        public int SendMinLimit { get; set; }
+        public decimal SendMinLimit { get; set; }
 
-        public int SendMaxLimit { get; set; }
+        public decimal SendMaxLimit { get; set; }
 
-        public int ReceiveMinLimit { get; set; }
+        public decimal ReceiveMinLimit { get; set; }
 
-        public int ReceiveMaxLimit { get; set; }
+        public decimal ReceiveMaxLimit { get; set; }
 
         public double Rate { get; set; }
